Update view edit

+++ resources/views/patients/show/show_allergy.blade.php
@@ -1,4 +1,5 @@
-<table class="table datatables" width="100%" cellspacing="0">
+<table class="table " width="100%" cellspacing="0">
+    <!-- class="datatables" -->
     <thead class="thead-light">
         <tr>
             <th>Alergia</th>
@@ -13,9 +14,7 @@
             <td>{!! $item->nombre_alergia !!}</td>
             <td class="text-right">
                 @if ( in_array(Auth::user()->role_id,[1,3]) )
-                <a href="/deleteallergy/{{$item->id}}" data-token="{{csrf_token()}}" class='btn btn-danger btn-xs'
-                    data-confirm="Seguro que quieres eliminar este pia?"
-                    onclick="return confirm('¿Estas segura que quieres el documento?')">
+                <a href="/deleteallergy/{{$item->id}}" data-token="{{csrf_token()}}" class='btn btn-danger btn-xs' data-confirm="Seguro que quieres eliminar este pia?" onclick="return confirm('¿Estas segura que quieres el documento?')">
                     <i class="fas fa-trash"></i>
                 </a>
                 @endif
