+++ resources/views/patients/show/show_past.blade.php
@@ -1,5 +1,6 @@
-<table class="table datatables dark" width="100%" cellspacing="0">
-    <thead class="thead-dark">
+<table class="table" width="100%" cellspacing="0">
+    <!-- class="datatables" -->
+    <thead class="thead-light">
         <tr>
             <th>Antecedentes</th>
             <th>Año</th>
@@ -17,11 +18,10 @@
             <td>{!! $patientPast->antecedentes_obs !!}</td>
             <td>
                 <div class='btn-group'>
-                    <a href="/pastdelete/{{$patientPast->id}}" data-token="{{csrf_token()}}"
-                        class='btn btn-block btn-danger btn-xs' data-confirm="Seguro que quieres eliminar este pia?"
-                        onclick="return confirm('¿Estas segura que quieres eliminar el antecedente?')">
+                    <a href="/pastdelete/{{$patientPast->id}}" data-token="{{csrf_token()}}" class='btn btn-block btn-danger btn-xs' data-confirm="Seguro que quieres eliminar este pia?" onclick="return confirm('¿Estas segura que quieres eliminar el antecedente?')">
                         <i class="fas fa-trash"></i> Eliminar
-                    </a> </div>
+                    </a>
+                </div>
             </td>
         </tr>
         @endforeach
